+++ backend/src/main/java/ca/metricalsky/winston/dto/fetch/FetchReplies.java
@@ -1,0 +1,10 @@
+package ca.metricalsky.winston.dto.fetch;
+
+import lombok.Data;
+
+@Data
+public class FetchReplies {
+
+    private String videoId;
+    private String commentId;
+}