+++ backend/src/main/java/ca/metricalsky/winston/mapper/entity/FetchRequestMapper.java
@@ -2,6 +2,7 @@ package ca.metricalsky.winston.mapper.entity;
 
 import ca.metricalsky.winston.dto.fetch.FetchChannel;
 import ca.metricalsky.winston.dto.fetch.FetchComments;
+import ca.metricalsky.winston.dto.fetch.FetchReplies;
 import ca.metricalsky.winston.dto.fetch.FetchRequestDto;
 import ca.metricalsky.winston.dto.fetch.FetchVideos;
 import ca.metricalsky.winston.entity.fetch.FetchRequest;
@@ -29,6 +30,9 @@ public class FetchRequestMapper {
         }
         if (fetchRequestDto.getComments() != null) {
             return commentsRequest(fetchRequestDto.getComments());
+        }
+        if (fetchRequestDto.getReplies() != null) {
+            return repliesRequest(fetchRequestDto.getReplies());
         }
         throw new AppException(HttpStatus.BAD_REQUEST, "The request is syntactically invalid and cannot be processed.");
     }
@@ -63,6 +67,13 @@ public class FetchRequestMapper {
         var fetchRequest = new FetchRequest();
         fetchRequest.setFetchType(FetchType.COMMENTS);
         fetchRequest.setObjectId(fetchComments.getVideoId());
+        return fetchRequest;
+    }
+
+    private FetchRequest repliesRequest(FetchReplies fetchReplies) {
+        var fetchRequest = new FetchRequest();
+        fetchRequest.setFetchType(FetchType.REPLIES);
+        fetchRequest.setObjectId(fetchReplies.getCommentId());
         return fetchRequest;
     }
 }