+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- New fetch operation to fetch replies by parent comment ID. (mddburgess/winston#8)
+
 ## [1.1.0] — 2025-04-05
 
 ### Added